Add files via upload
This is the first Qiskit implementation file in my Quantum Journey.  This starts with the basics of Quantum Gates of the Qiskit and goes into advanced, up to Variational Quantum Gates. I gave a description (explanation) for each step that was included in it. I hope it serves as a useful resource and helps the Quantum Community